# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMC Documentation",
        "description": "Api para criaçao e gerenciamento de prontuarios, juntamente com analise de batimentos cardiacos",
        "contact": {
            "url": "https://github.com/gmatheus66"
        },
        "license": {
            "name": "License: MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apisimc.herokuapp.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/doctor/detail": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "description": "Returns a detail doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/register": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "description": "Doctor Register",
                "operationId": "App\\Http\\Controllers\\DoctorController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "crm",
                        "in": "query",
                        "description": "Crm",
                        "required": true
                    },
                    {
                        "name": "specialization",
                        "in": "query",
                        "description": "Specialization",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/logout": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "description": "Login Doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/login": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "description": "Login Doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/check": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "description": "Returns a check Doctor",
                "operationId": "App\\Http\\Controllers\\DoctorController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/patient/detail": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns a detail Patient",
                "operationId": "App\\Http\\Controllers\\PatientController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/patient/check": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns a check Patient",
                "operationId": "App\\Http\\Controllers\\PatientController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/relative/detail": {
            "get": {
                "tags": [
                    "Relative"
                ],
                "description": "Returns a detail Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::detail_auth_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/relative/check": {
            "get": {
                "tags": [
                    "Relative"
                ],
                "description": "Returns a check Relative",
                "operationId": "App\\Http\\Controllers\\RelativeController::check_user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}